backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.0.7
openai==1.3.7
cohere==4.39.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
